refactoring outline
