+++ debvader/debvader.py
@@ -1,166 +1,13 @@
-import os
-
 import numpy as np
 import pandas as pd
-import pkg_resources
 import scipy
-import sep
 import tensorflow as tf
-from scipy import optimize
-
-from debvader import model
+from object_detection import detect_objects, extract_cutouts
+from position_optimization import position_optimization
 
 
 def mse(img1, img2):
     return np.mean(np.square(img1 - img2))
-
-
-def extract_cutouts(
-    field_image, field_size, galaxy_distances_to_center, cutout_size=59, nb_of_bands=6
-):
-    """
-    Extract the cutouts around particular galaxies in the field
-    parameters:
-        field_image: image of the field to deblend
-        field_size: size of the field
-        galaxy_distances_to_center: distances of the galaxies to deblend from the center of the field. In pixels.
-        cutout_size: size of the stamps
-    """
-    cutout_images = np.zeros(
-        (len(galaxy_distances_to_center), cutout_size, cutout_size, nb_of_bands)
-    )
-    list_idx = []
-    flag = False
-
-    for i in range(len(galaxy_distances_to_center)):
-        try:
-            x_shift = galaxy_distances_to_center[i][0]
-            y_shift = galaxy_distances_to_center[i][1]
-
-            x_start = -int(cutout_size / 2) + int(x_shift) + int(field_size / 2)
-            x_end = int(cutout_size / 2) + int(x_shift) + int(field_size / 2) + 1
-
-            y_start = -int(cutout_size / 2) + int(y_shift) + int(field_size / 2)
-            y_end = int(cutout_size / 2) + int(y_shift) + int(field_size / 2) + 1
-
-            cutout_images[i] = field_image[0, x_start:x_end, y_start:y_end]
-            list_idx.append(i)
-
-        except ValueError:
-            flag = True
-
-    if flag:
-        print(
-            "Some galaxies are too close from the border of the field to be considered here."
-        )
-
-    return cutout_images, list_idx
-
-
-def detect_objects(field_image):
-    """
-    Detect the objects in the field_image image using the SExtractor detection algorithm.
-    test for dev branch
-    """
-    field_image = field_image.copy()
-    field_size = field_image.shape[1]
-    galaxy_distances_to_center = []
-
-    r_band_data = field_image[0, :, :, 2].copy()
-    bkg = sep.Background(r_band_data)
-
-    r_band_foreground = r_band_data - bkg
-
-    DETECT_THRESH = 0.8
-    deblend_cont = 0.00001
-    deblend_nthresh = 64
-    minarea = 4
-    filter_type = "conv"
-    # 7x7 convolution mask of a gaussian PSF with FWHM = 3.0 pixels.
-    filter_kernel = np.array(
-        [
-            [0.004963, 0.021388, 0.051328, 0.068707, 0.051328, 0.021388, 0.004963],
-            [0.021388, 0.092163, 0.221178, 0.296069, 0.221178, 0.092163, 0.021388],
-            [0.051328, 0.221178, 0.530797, 0.710525, 0.530797, 0.221178, 0.051328],
-            [0.068707, 0.296069, 0.710525, 0.951108, 0.710525, 0.296069, 0.068707],
-            [0.051328, 0.221178, 0.530797, 0.710525, 0.530797, 0.221178, 0.051328],
-            [0.021388, 0.092163, 0.221178, 0.296069, 0.221178, 0.092163, 0.021388],
-            [0.004963, 0.021388, 0.051328, 0.068707, 0.051328, 0.021388, 0.004963],
-        ]
-    )
-
-    objects = sep.extract(
-        data=r_band_foreground,
-        thresh=DETECT_THRESH,
-        err=bkg.globalrms,
-        deblend_cont=deblend_cont,
-        deblend_nthresh=deblend_nthresh,
-        minarea=minarea,
-        filter_kernel=filter_kernel,
-        filter_type=filter_type,
-    )
-
-    for i in range(len(objects["y"])):
-        galaxy_distances_to_center.append(
-            (
-                np.round(-int(field_size / 2) + objects["y"][i]),
-                np.round(-int(field_size / 2) + objects["x"][i]),
-            )
-        )
-
-    return np.array(galaxy_distances_to_center)
-
-
-def load_deblender(
-    survey, input_shape, latent_dim, filters, kernels, return_encoder_decoder_z=False
-):
-    """
-    load weights trained for a particular dataset
-    parameters:
-        survey: string calling the particular dataset (choices are: "dc2")
-        input_shape: shape of input tensor
-        latent_dim: size of the latent space
-        filters: filters used for the convolutional layers
-        kernels: kernels used for the convolutional layers
-    """
-    # Create the model
-    net, encoder, decoder, z = model.create_model_vae(
-        input_shape,
-        latent_dim,
-        filters,
-        kernels,
-        conv_activation=None,
-        dense_activation=None,
-    )
-
-    # Define the loss function
-    def vae_loss(x, x_decoded_mean):
-        # xent_loss = K.mean(
-        #    K.sum(K.binary_crossentropy(x, x_decoded_mean), axis=[1, 2, 3])
-        # )
-        return -x_decoded_mean.log_prob(x)  # xent_loss
-
-    # Set the decoder as non-trainable
-    decoder.trainable = False
-
-    # Compile the model
-    net.compile(
-        optimizer=tf.optimizers.Adam(learning_rate=1e-4),
-        loss=vae_loss,
-        experimental_run_tf_function=False,
-    )
-
-    # Load the weights corresponding to the chosen survey
-    data_path = pkg_resources.resource_filename("debvader", "data/")
-    loading_path = os.path.join(data_path, "weights/", survey, "not_normalised/loss/")
-    print(loading_path)
-    latest = tf.train.latest_checkpoint(loading_path)
-    net.load_weights(latest)
-
-    if return_encoder_decoder_z:
-        return net, encoder, decoder, z
-    else:
-        return net
 
 
 def deblend(net, images, normalised=False):
@@ -180,55 +27,6 @@ def deblend(net, images, normalised=False):
     return net(tf.cast(images, tf.float32)).mean().numpy(), net(
         tf.cast(images, tf.float32)
     )
-
-
-def position_optimization(
-    field_image,
-    output_image_mean_padded,
-    galaxy_distance_to_center,
-    method="scipy-minimize",
-):
-    """
-    Find shifts in the position of the deblended galaxy to minimize the mse between field_image
-
-    parameters:
-        field image: image of the entire field of galaxy to be deblended.
-        output_images_mean_padded: predicted image of the galaxy that is to be optimized.
-        galaxy_distances_to_center: distance of the predicted galaxy from the center, as detected by the detection algorithm.
-    """
-
-    assert method in ["scipy-minimize"]
-
-    if method == "scipy-minimize":
-
-        def fun(x, img, net_output):
-            """
-            parameters:
-                x: shifts for x and y position
-                img: field image
-                net_output: predicted image if the galaxy
-            """
-            return mse(img, scipy.ndimage.shift(net_output, shift=(x[0], x[1])))
-
-        r_band_field = field_image[:, :, 2]
-        r_band_perdiction = output_image_mean_padded[:, :, 2]
-        opt = optimize.least_squares(
-            fun,
-            (0.0, 0.0),
-            args=(
-                r_band_field,
-                scipy.ndimage.shift(
-                    r_band_perdiction,
-                    shift=(galaxy_distance_to_center[0], galaxy_distance_to_center[1]),
-                ),
-            ),
-            bounds=(-3, 3),
-        )
-
-        shift_x = opt.x[0]
-        shift_y = opt.x[1]
-
-    return shift_x, shift_y
 
 
 class DeblendField:
